+++ dev-demos/gallery/scale/cat.tsx
@@ -1,5 +1,5 @@
 import { PolygonLayer, Scene } from '@antv/l7';
-import { Mapbox } from '@antv/l7-maps';
+import { Map } from '@antv/l7-maps';
 import React, { useEffect } from 'react';
 import { useData, addLayers } from './useLine';
 
@@ -9,7 +9,7 @@ export default () => {
   useEffect(() => {
     const scene = new Scene({
       id: 'map',
-      map: new Mapbox({
+      map: new Map({
         pitch: 0,
         style: 'light',
         center: [-96, 37.8],
