+++ package.json
@@ -44,6 +44,7 @@
     "@types/gl": "^4.1.0",
     "@types/hoist-non-react-statics": "^3.3.1",
     "@types/jest": "^25.2.1",
+    "@types/leaflet": "^1.7.11",
     "@types/node": "13.11.1",
     "@types/offscreencanvas": "^2019.7.0",
     "@types/react-router-dom": "^5.3.2",
@@ -101,6 +102,7 @@
     "jest": "^24.9.0",
     "jest-canvas-mock": "^2.4.0",
     "jest-styled-components": "^6.2.1",
+    "leaflet": "^1.8.0",
     "lerc": "^3.0.0",
     "lerna": "^3.16.4",
     "lint-staged": "^9.2.4",
