+++ packages/core/src/services/scene/SceneService.ts
@@ -1,5 +1,5 @@
 // @ts-ignore
-import { AsyncParallelHook } from '@antv/async-hook';
+import { AsyncSeriesHook  } from '@antv/async-hook';
 import { $window, DOM } from '@antv/l7-utils';
 import elementResizeEvent, { unbind } from 'element-resize-event';
 import { EventEmitter } from 'eventemitter3';
@@ -106,7 +106,7 @@ export default class Scene extends EventEmitter implements ISceneService {
   private markerContainer: HTMLElement;
 
   private hooks: {
-    init: AsyncParallelHook;
+    init: AsyncSeriesHook;
   };
 
   public constructor() {
@@ -119,7 +119,7 @@ export default class Scene extends EventEmitter implements ISceneService {
        * 2. initRenderer：初始化渲染引擎
        * 3. initWorker：初始化 Worker
        */
-      init: new AsyncParallelHook(),
+      init: new AsyncSeriesHook(),
     };
   }
 
@@ -144,31 +144,15 @@ export default class Scene extends EventEmitter implements ISceneService {
         this.map.onCameraChanged((viewport: IViewport) => {
           this.cameraService.init();
           this.cameraService.update(viewport);
-          if (this.map.version !== 'GAODE2.x') {
-            // not amap2
-            resolve();
-          }
-        });
-
-        if (this.map.version !== 'GAODE2.x') {
-          // not amap2
-          this.map.init();
-        } else {
-          // amap2
           resolve();
-        }
+        });
+        this.map.init();
       });
-
-      if (this.map.version === 'GAODE2.x' && this.map.initViewPort) {
-        // amap2
-        await this.map.init();
-        this.map.initViewPort();
-      }
-
+    
+      
       // 重新绑定非首次相机更新事件
       this.map.onCameraChanged(this.handleMapCameraChanged);
       this.map.addMarkerContainer();
-
       // 初始化未加载的marker;
       this.markerService.addMarkers();
       this.markerService.addMarkerLayers();
@@ -185,6 +169,7 @@ export default class Scene extends EventEmitter implements ISceneService {
      * 初始化渲染引擎
      */
     this.hooks.init.tapPromise('initRenderer', async () => {
+      console.log(this.map)
       // 创建底图之上的 container
       const $container = createRendererContainer(
         this.configService.getSceneConfig(this.id).id || '',
@@ -321,7 +306,7 @@ export default class Scene extends EventEmitter implements ISceneService {
     if (!this.inited) {
       // 还未初始化完成需要等待
 
-      await this.initPromise;
+      await this.initPromise; // 初始化地图和渲染
       if (this.destroyed) {
         this.destroy();
       }
@@ -336,6 +321,7 @@ export default class Scene extends EventEmitter implements ISceneService {
           console.warn(e);
         }
       }
+
       // FIXME: 初始化 marker 容器，可以放到 map 初始化方法中？
       this.layerService.initLayers();
       this.controlService.addControls();
