+++ packages/core/typings/l7-tiny-sdf.d.ts
@@ -1,3 +1,2 @@
 declare module 'l7hammerjs';
 declare module 'l7-tiny-sdf';
-declare module '@antv/l7-maps';
