+++ packages/maps/src/amap/index.ts
@@ -1,5 +1,5 @@
 import { IAMapInstance } from '../../typings/index';
-import BaseMapWrapper from '../BaseMapWrapper';
+import BaseMapWrapper from '../utils/BaseMapWrapper';
 import AMapService from './map';
 
 export default class AMapWrapper extends BaseMapWrapper<
