+++ packages/maps/src/map/index.ts
@@ -1,5 +1,5 @@
 import { Map } from '@antv/l7-map';
-import BaseMapWrapper from '../BaseMapWrapper';
+import BaseMapWrapper from '../utils/BaseMapWrapper';
 import MapService from './map';
 export default class MapboxWrapper extends BaseMapWrapper<Map> {
   protected getServiceConstructor() {
