+++ packages/maps/src/mapbox/index.ts
@@ -1,6 +1,6 @@
 import { Map } from 'mapbox-gl';
 import { IMapboxInstance } from '../../typings/index';
-import BaseMapWrapper from '../BaseMapWrapper';
+import BaseMapWrapper from '../utils/BaseMapWrapper';
 import './logo.css';
 import MapboxService from './map';
 export default class MapboxWrapper extends BaseMapWrapper<
