+++ packages/maps/src/utils/Viewport.ts
@@ -2,7 +2,7 @@ import { IMapCamera, IViewport } from '@antv/l7-core';
 import WebMercatorViewport from 'viewport-mercator-project';
 
 export default class Viewport implements IViewport {
-  private viewport: WebMercatorViewport;
+  public viewport: WebMercatorViewport;
 
   public syncWithMapCamera(mapCamera: Partial<IMapCamera>) {
     const {
@@ -14,11 +14,24 @@ export default class Viewport implements IViewport {
       viewportWidth,
     } = mapCamera;
 
+    const preView = this.viewport
+      ? {
+          width: this.viewport.width,
+          height: this.viewport.height,
+          longitude: this.viewport.center[0],
+          latitude: this.viewport.center[1],
+          zoom: this.viewport.zoom,
+          pitch: this.viewport.pitch,
+          bearing: this.viewport.bearing,
+        }
+      : {};
+
     /**
      * Deck.gl 使用的也是 Mapbox 同步相机，相机参数保持一致
      * 例如相机高度固定为 height * 1.5，因此不需要传
      */
     this.viewport = new WebMercatorViewport({
+      ...preView,
       width: viewportWidth,
       height: viewportHeight,
       longitude: center && center[0],
