+++ packages/maps/src/utils/utils.ts
@@ -42,5 +42,5 @@ export function toPaddingOptions(padding: IPadding = {}) {
     }
   }
 
-  return Object.assign({},defaultPadding,padding)
+  return Object.assign({}, defaultPadding, padding);
 }
