+++ packages/source/package.json
@@ -24,7 +24,7 @@
   "author": "lzxue",
   "license": "ISC",
   "dependencies": {
-    "@antv/async-hook": "^2.1.0",
+       "@antv/async-hook": "^2.2.2",
     "@antv/l7-core": "2.9.25",
     "@antv/l7-utils": "2.9.25",
     "@babel/runtime": "^7.7.7",
