+++ tsconfig.json
@@ -31,6 +31,7 @@
       "@antv/l7-utils": ["packages/utils/src"],
       "@antv/l7-test-utils":["packages/test-utils/src"],
       "@antv/l7": ["packages/l7/src"],
+      "@antv/l7-leaflet": ["packages/leaflet/src"],
       "*": ["packages", "typings/*"]
     }
   },
